Reference

Contents

Reference

Contents

Exceptions

Context Handling

Stanza

Connection Handling

Stream-Management

Stanza Handlers, Timed Event Handlers

Logging

Index

Exceptions

Strophe.StropheError — Type

Handle errors related to libstrophe.

Context Handling

Strophe.Context — Type

Context()

A wrapper for the internal libstrophe context. It handles allocation and de-allocation for you, as well as translating the internal libstrophe logger to Julia's native logging system.

Strophe.DEFAULT_CONTEXT — Constant

The default context provided by the library.

Strophe.context — Function

context([obj])

Get the pointer to the LibStrophe.xmpp_ctx_t object associated to obj. If called without argument, return the default context.

Strophe.run — Function

run([obj])

Run the internal libstrophe main loop in the context related to obj (or the default context is none is provided) until it is stopped by a stop call. This is not directly wrapping the internal LibStrophe.xmpp_run function. Instead, it is a nearly-identical rewrite that introduces a yield() call after each LibStrophe.xmpp_run_once call, allowing other tasks to be run.

Warn

This does not make the library thread-safe, and all Strophe-related things should run in the same thread. It does, however, allow you to perform other tasks in parallel.

See also run_once, timeout!, context, LibStrophe.xmpp_run.

Strophe.run_once — Function

run_once([obj,] timeout)

Run one iteration of the internal libstrophe main loop in the context related to obj, or the default context if none is provided. Will send queued data, and go through the event loop once. It will wait no more than timeout milliseconds.

See also run, LibStrophe.xmpp_run_once.

Strophe.stop — Function

stop([obj])

Stop the main loop triggered by run, using the default context by default.

See also run, LibStrophe.xmpp_stop.

Strophe.timeout! — Function

timeout!([obj, ], timeout)

Set the timeout in milliseconds to use when calling run. Use the default context by default.

See also run, LibStrophe.xmpp_ctx_set_timeout.

Stanza

Strophe.Stanza — Type

Wraps a pointer to a libstrophe stanza, handling release at finalization.

Constructors

Stanza()
Stanza(ctx)
Stanza(ctx; text)
Stanza(ctx; name, attributes...)

Construct a stanza by extracting the context of ctx using context. Defaults to using the default context. The name keyword will make constructor call name! function on the stanza and make it a tag stanza. The other keyword arguments attributes. should be of type AbstractString and will be passed to setindex!, and require the name keyword argument. The text keyword will make the constructor call the text! function on the stanza and make it a text stanza. This keyword arguments thus excludes the use of others.

Stanza([ctx,] str::String)

Construct a stanza from the given string. Will throw a StropheError on failure. Defaults to using the default context. See also LibStrophe.xmpp_stanza_new_from_string.

Stanza(stanza::Stanza)

Copy an existing stanza along with all its children and returns the new stanza and children with a reference count of 1. The returned stanza will have no parent and no siblings. This is useful for extracting a child stanza for inclusion in another tree. See also LibStrophe.xmpp_stanza_copy.

See also message, reply, iq, presence, reply_error, stream_error for specialized constructors.

Callable API

Stanza object are callable. This allows you to build stanzas with children in a compact way. The return value of calling a Stanza object is the object itself. In the following, we assume s is a Stanza object.

s(text)

If s is not a tag stanza, set the text of a stanza (making it a text stanza). Otherwise, add a text child containing text.

s(children...[; attributes])

When s is a tag stanza, add children to the stanza, and optionally add the relevant attributes.

Examples

Here are two ways of creating the <iq/> following stanza (inspired by the bot example):

<iq
    type='result'
    to='romeo@example.net'
    from='juliet@example.com'
    id='version_1'>
  <query xmlns='jabber:iq:version'>
    <name>LibStrophe example bot</name>
    <version>1.0</version>
  </query>
</iq>

First manually:

Stanza(name="iq", type="result",
       to="romeo@example.net", from="juliet@example.com",
       id="version_1")(
    Stanza(name="query", xmlns="jabber:iq:version")(
        Stanza(name="name")("LibStrophe example bot"),
        Stanza(name="version")("1.0")
    )
)

Alternatively, using convenience constructors and functions:

iq(type="result", id="version_1")(
    Stanza(name="query", xmlns="jabber:iq:version")(
        Stanza(name="name")("LibStrophe example bot"),
        Stanza(name="version")("1.0")
    )
)

Note that you can also build the stanza bits by bits:

s = iq(type="result")
id!(s, "version_1")
to!(s, "romeo@example.net")
from!(s, "juliet@example.com")
query = Stanza()
name!(query, "query")
ns!(query, "jabber:iq:version")
name = Stanza(name="name")("LibStrophe example bot")
version = Stanza(name="version")("1.0")
child!(query, name)
child!(query, version)
child!(iq, query)

Note

Unlike when using the low-level API, you do not need to take care of memory management manually here.

See also iq, id!, to!, from!, name!, ns!, child!.

Strophe.stanza — Function

stanza(obj)

Get the pointer to the LibStrophe.xmpp_stanza_t object associated to obj.

Strophe.clone! — Function

clone!(stanza)

Clone a stanza. The returned Stanza object will point to the same LibStrophe.xmpp_stanza_t object, with its reference counter increased by one.

See also LibStrophe.xmpp_stanza_clone.

Strophe.release! — Function

release!(stanza)

Release a stanza object and all of its children. This will cause the stanza's context to decrement the stanza's counter, and will lead to the object(s) to be freed if the counter falls to zero. In this case, accessing the stanza is no longer valid. This is typically used by the finalizer of Stanza.

Strophe.is_text — Function

is_text(stanza)

Return true if the stanza is a text node.

Strophe.is_tag — Function

is_tag(stanza)

Return true if the stanza is a tag node.

Strophe.render_to_string — Function

render_to_string(stanza)

Render a stanza to text.

Strophe.children — Function

children(stanza)

Iterate over the children of a stanza.

See also LibStrophe.xmpp_stanza_get_children.

Strophe.siblings — Function

siblings(stanza)

Iterate over the next siblings of stanza.

See also LibStrophe.xmpp_stanza_get_next.

Strophe.child — Function

child(stanza, name)

Get the first child of stanza named name, or return nothing if none is found.

See also LibStrophe.xmpp_stanza_get_child_by_name.

child(stanza, name, ns)

Get the first child of stanza named name and in namespace ns, or return nothing if none is found.

See also LibStrophe.xmpp_stanza_get_child_by_name_and_ns.

child(stanza; name, ns)

Get the first child of stanza named name and in namespace ns, or return nothing if none is found. Both keywords can be specified, and at least one must be specified.

See also LibStrophe.xmpp_stanza_get_child_by_name, LibStrophe.xmpp_stanza_get_child_by_ns, LibStrophe.xmpp_stanza_get_child_by_name_and_ns.

Strophe.child! — Function

child!(stanza, child[; do_clone=true])

Add to stanza the child child. if do_clone is set to false, the user releases ownership of the child and must not release it or transfer ownership.

Will throw a StropheError when adding the child fails.

See also LibStrophe.xmpp_stanza_add_child, LibStrophe.xmpp_stanza_add_child_ex.

Base.getindex — Method

getindex(stanza, attribute)

Get an attribute from a stanza. Throws a KeyError if the attribute does not exist.

See also LibStrophe.xmpp_stanza_get_attribute.

Base.keys — Method

keys(stanza)

Return the list of attributes defined for the stanza.

See also LibStrophe.xmpp_stanza_get_attribute_count, LibStrophe.xmpp_stanza_get_attributes, Base.pairs.

Base.values — Method

values(stanza)

Return the list of attributes defined for the stanza.

See also LibStrophe.xmpp_stanza_get_attribute_count, LibStrophe.xmpp_stanza_get_attributes, Base.pairs.

Base.pairs — Method

pairs(stanza)

Return the an array of attribute key=>value pairs.

See also LibStrophe.xmpp_stanza_get_attribute_count, LibStrophe.xmpp_stanza_get_attributes, Base.pairs.

Strophe.text — Function

text(stanza)

Get the text data for a text stanza.

See also LibStrophe.xmpp_stanza_get_text, text!.

Strophe.text! — Function

text!(stanza, text)

Set the text data for a text stanza. This will throw a StropheError when ran on a stanza that has a name.

See also text, LibStrophe.xmpp_stanza_set_text.

Strophe.name — Function

name(stanza)

Get the stanza name.

See also LibStrophe.xmpp_stanza_get_name, name!.

Strophe.name! — Function

name!(stanza, name)

Set the name of a stanza. Throws a StropheError on failure.

See also LibStrophe.xmpp_stanza_set_name, name.

Base.setindex! — Method

setindex!(stanza, value, key)

Set a stanza attribute key to value. Throws a StropheError on failure.

See also LibStrophe.xmpp_stanza_set_attribute.

Base.delete! — Method

delete!(stanza, attribute)

Delete the attribute of stanza.

See also LibStrophe.xmpp_stanza_del_attribute.

Strophe.ns — Function

ns(obj)

Get the namespace attribute of stanza. Return the string with the 'xmlns' attribute value, or nothing if it does not exist.

See also ns!, LibStrophe.xmpp_stanza_get_ns.

Strophe.ns! — Function

ns!(obj, ns)

Set the namespace attribute of stanza.

See also ns, LibStrophe.xmpp_stanza_set_ns.

Strophe.type — Function

type(obj)

Get the type attribute of stanza. Return the string with the 'type' attribute value, or nothing if it does not exist. .

See also type!, LibStrophe.xmpp_stanza_get_type.

Strophe.type! — Function

type!(obj, type)

Set the type attribute of stanza.

See also type, LibStrophe.xmpp_stanza_set_type.

Strophe.id — Function

id(obj)

Get the id attribute of stanza. Return the string with the 'id' attribute value, or nothing if it does not exist.

See also id!, LibStrophe.xmpp_stanza_get_id.

Strophe.id! — Function

id!(obj, id)

Set the id attribute of stanza.

See also id, LibStrophe.xmpp_stanza_set_id.

Strophe.to — Function

to(obj)

Get the to attribute of stanza. Return the string with the 'to' attribute value, or nothing if it does not exist.

See also to!, LibStrophe.xmpp_stanza_get_to.

Strophe.to! — Function

to!(obj, to)

Set the to attribute of stanza.

See also to, LibStrophe.xmpp_stanza_set_to.

Strophe.from — Function

from(obj)

Get the from attribute of stanza. Return the string with the 'from' attribute value, or nothing if it does not exist.

See also from!, LibStrophe.xmpp_stanza_get_from.

Strophe.from! — Function

from!(obj, from)

Set the from attribute of stanza.

See also from, LibStrophe.xmpp_stanza_set_from.

Strophe.reply — Function

reply(obj)

Create a stanza object in reply to another.

See also LibStrophe.xmpp_stanza_reply.

Strophe.reply_error — Function

reply_error(stanza, error_type, condition[, text])

Create an error stanza object in reply to the provided stanza.

Check https://tools.ietf.org/html/rfc6120#section-8.3 for details.

Parameters:

stanza the stanza to reply to

error_type the type attribute for the <error/> child element

condition the defined-condition (e.g. "item-not-found")

text optional description

See also LibStrophe.xmpp_stanza_reply_error.

Strophe.message — Function

message([ctx]; type, to, id)

Create a <message/> stanza object with the given attributes. Attributes are optional. Defaults to the default context.

See also LibStrophe.xmpp_message_new.

Strophe.body — Function

body(message)

Get text from <body/> child element. Will return nothing when message does not have a <body/> element and on memory allocation error.

See also LibStrophe.xmpp_message_get_body, body!.

Strophe.body! — Function

body!(message, text)

Add a <body/> child element to a <message/> stanza with the given text. Will throw a StropheError on failure.

See also LibStrophe.xmpp_message_set_body, body.

Strophe.iq — Function

iq([ctx]; type, id)

Create a <iq/> stanza object with given attributes. Attributes are optional. Defaults to the default context.

See also LibStrophe.xmpp_iq_new.

Strophe.presence — Function

presence([ctx])

Create a <presence/> stanza object. Defaults to the default context.

See also LibStrophe.xmpp_presence_new.

Strophe.stream_error — Function

stream_error([ctx]; type, text)

Create an <stream:error/> stanza object with given type and an optional error text. Defaults to the default context.

See also LibStrophe.xmpp_error_new

Connection Handling

Strophe.ParametrizedConnection — Type

ParametrizedConnection(conn::ParametrizedConnection)
ParametrizedConnection(; type::LibStrophe.xmpp_conn_type_t, conn::Union{Nothing, Connection} = nothing, ctx::Context = context(), host = "", port = 0, handler = default_connection_handler, stream_management_callback=nothing)

A wrapper around a Connection to remember the kind of connection, the host to connect to, the port, and the handler.

type::Strophe.LibStrophe.xmpp_conn_type_t: Connection type (component or client).

conn::Strophe.Connection: Internal Connection object.

host::String: Default host.

port::Int32: Default port.

handler::FunctionWrappers.FunctionWrapper{Nothing, Tuple{Strophe.ParametrizedConnection, Strophe.LibStrophe.xmpp_conn_event_t, Int32, Ptr{Strophe.LibStrophe.xmpp_stream_error_t}}}: Handler function for connection events.

stream_management_callback::Union{Nothing, Ptr{Nothing}}: Stream management callback.

See also ClientConnection, ComponentConnection, connect, disconnect.

Strophe.ClientConnection — Function

ClientConnection(; kwargs...)

Build a ParametrizedConnection for a client connection (type set to LibStrophe.XMPP_CLIENT.

See also ParametrizedConnection, ComponentConnection.

ClientConnection(conn)

Clone an existing connection to create a client connection.

See also LibStrophe.xmpp_conn_clone.

Strophe.ComponentConnection — Function

ComponentConnection(; kwargs...)

Build a ParametrizedConnection for a client connection (type set to LibStrophe.XMPP_COMPONENT.

See also ParametrizedConnection, ClientConnection.

ComponentConnection(conn)

Clone an existing connection to create a component connection.

See also LibStrophe.xmpp_conn_clone.

Strophe.connect — Function

connect(conn)

Initiate the connection parametrized by conn.

See also disconnect, ParametrizedConnection, LibStrophe.xmpp_connect_client, LibStrophe.xmpp_connect_component.

Strophe.disconnect — Function

disconnect(conn)

Initiate the connection parametrized by conn.

See also connect, ParametrizedConnection.

Strophe.send — Function

send(conn, stanza)

Send stanza through conn. Will use connection and stanza to get down to the unerlying libstrophe pointers.

Strophe.Connection — Type

Connection()
Connection(context::Context)
Connection(connection::Connection)

A wrapper for a libstrophe connection. It handles allocation and de-allocation for you. If called with an existing Connection, use the internal LibStrophe.xmpp_conn_clone to clone the connection. Will use the default context if none is provided.

See also Context.

Strophe.connection — Function

connection(obj)

Get the pointer to the LibStrophe.xmpp_conn_t associated to obj.

Strophe.jid — Function

jid(obj)

Get the JID which is or will be bound to the connection associated to obj.

See also connection, LibStrophe.xmpp_conn_get_jid.

Strophe.bound_jid — Function

bound_jid(obj)

Get the JID discovered during binding time for the connection associated to obj.

See also connection, LibStrophe.xmpp_conn_get_bound_jid.

Strophe.jid! — Function

jid!(obj, jid)

Set the JID of the user that will be bound to the connection associated to obj.

See also connection, LibStrophe.xmpp_conn_set_jid.

Strophe.pass — Function

pass(obj)

Get the password used for authentication of the connection associated to obj.

See also connection, LibStrophe.xmpp_conn_get_pass.

Strophe.pass! — Function

pass!(connection, password)

Set the password used to authenticate the connection associated to obj.

See also connection, LibStrophe.xmpp_conn_set_pass.

Strophe.flags — Function

flags(obj)

Get the flags of the connection associated to obj.

See also connection, LibStrophe.xmpp_conn_get_flags, flags!.

Strophe.flags! — Function

flags!(obj, flags)

Set the flags associated to the connection of obj.

See also connection, LibStrophe.xmpp_conn_set_flags, flags.

Strophe.flagenabled — Function

flagenabled(obj, flag)

Return true if flag is set in the connection associated to obj.

See also connection, flags.

Strophe.enableflag! — Function

enableflag!(obj, flag)

Enable flag in the connection associated to obj.

See also connection, flags!.

Strophe.disabletls — Function

disabletls(obj)

Return true if the DISABLE_TLS flag is set in the connection associated to obj.

See also connection, flags, flagenabled.

Strophe.disabletls! — Function

disabletls!(obj, st=true)

Set the DISABLE_TLS flag of connection associated to obj to st.

See also connection, flags!, enableflag!.

Strophe.mandatorytls — Function

mandatorytls(obj)

Return true if the MANDATORY_TLS flag is set in the connection associated to obj.

See also connection, flags, flagenabled.

Strophe.mandatorytls! — Function

mandatorytls!(obj, st=true)

Set the MANDATORY_TLS flag in the connection associated to obj to st.

See also connection, flags!, enableflag!.

Strophe.legacyssl — Function

legacyssl(obj)

Return true if the LEGACY_SSL flag is set in the connection associated to obj.

See also connection, flags, flagenabled.

Strophe.legacyssl! — Function

legacyssl!(obj, st=true)

Set the LEGACY_SSL flag in the connection associated to obj to st.

See also connection, flags!, enableflag!.

Strophe.trusttls — Function

trusttls(obj)

Return true if the TRUST_TLS flag is set in the connection associated to obj.

See also connection, flags, flagenabled.

Strophe.trusttls! — Function

trusttls!(obj, st=true)

Set the TRUST_TLS flag in the connection associated to obj to st.

See also connection, flags!, enableflag!.

Strophe.legacyauth — Function

legacyauth(obj)

Return true if the LEGACY_AUTH flag is set in the connection associated to obj.

See also connection, flags, flagenabled.

Strophe.legacyauth! — Function

legacyauth!(obj, st=true)

Set the LEGACY_AUTH flag in the connection associated to obj to st.

See also connection, flags!, enableflag!.

Strophe.disablesm — Function

disablesm(obj)

Return true if the DISABLE_SM flag is set in the connection associated to obj.

See also connection, flags, flagenabled.

Strophe.disablesm! — Function

disablesm!(obj, st=true)

Set the DISABLE_SM flag in the connection associated to obj to st.

See also connection, flags!, enableflag!.

Strophe.enablecompression — Function

enablecompression(obj)

Return true if the ENABLE_COMPRESSION flag is set in the connection associated to obj.

See also connection, flags, flagenabled.

Strophe.enablecompression! — Function

enablecompression!(obj, st=true)

Set the ENABLE_COMPRESSION flag in the connection associated to obj to st.

See also connection, flags!, enableflag!.

Strophe.compressiondontreset — Function

compressiondontreset(obj)

Return true if the COMPRESSION_DONT_RESET flag is set in the connection associated to obj.

See also connection, flags, flagenabled.

Strophe.compressiondontreset! — Function

compressiondontreset!(obj, st=true)

Set the COMPRESSION_DONT_RESET flag in the connection associated to obj to st.

See also connection, flags!, enableflag!.

Strophe.sockopt_cb_keepalive — Constant

A pointer to LibStrophe.xmpp_sockopt_cb_keepalive. See for example usage.

Stream-Management

LibStrophe implements XEP-0198, and these functionnalities are exposed in Strophe.jl.

Strophe.StreamManagementState — Type

Wraps a pointer to a libstrophe stream-management object.

This can be used to store the stream-management state of a connection after deconnecting and reloading it when connecting again.

See also XEP-0198.

Examples

Strophe.run() # After this call, the main loop ends and we disconnected
sm_state = Strophe.stream_management_state(conn)
# Later, we can reconnect and load this stream management state
Strophe.stream_management_state!(new_conn, sm_state)
Strophe.connect(new_conn)
Strophe.run()

Strophe.stream_management_state — Function

stream_management_state(conn)

Get the StreamManagementState associated to conn.

Strophe.stream_management_state! — Function

stream_management_state!(conn, sm_state)

Set the StreamManagementState associated to conn to sm_state. The sm_state object will be rendered unusable.

Strophe.restore_state_management_state! — Function

restore_state_management_state!(conn, sm_state)

Restore the serialized stream management state of a connection.

See the stream_management_callback field of ParametrizedConnection for setting a callback function that will be called to store the state.

Stanza Handlers, Timed Event Handlers

Strophe.add_handler! — Function

add_handler!(conn, handler[, args][; name][, namespace][,type])
add_handler!(conn, event, handler[, args])

Add a stanza handler.

The handler will be called when any of the specified filters match. The filters behave as follow:

name: matches to the top level stanza name,

namespace: matches the namespace (xmlns attribute) of either the top level stanza or any of it's immediate children. This allows you to to handle specific <iq/> stanzas based on the <query/> child namespace.

type: matches the type attribute of the top level stanza.

Handlers should accept two parameters: a connection (pointer to LibStrophe.xmpp_conn_t) and a stanza (pointer to LibStrophe.xmpp_stanza_t).

See also LibStrophe.xmpp_handler_add, delete_handler!.

Strophe.delete_handler! — Function

delete_handler!(conn, handler)

Delete a stanza handler.

See also LibStrophe.xmpp_handler_delete, add_handler!.

Strophe.REGISTERED_HANDLERS — Constant

Store the registered handlers. Keys are objectid of the wrapped functions.

Strophe.add_id_handler! — Function

add_id_handler!(conn, handler, id)

Add an id based stanza handler.

This function adds a stanza handler for an <iq/> stanza of type 'result' or 'error' with a specific id attribute. This can be used to handle responses to specific <iq/>s.

If the handler function returns true, it will be kept, and if it returns false, it will be deleted from the list of handlers.

Handlers should accept two parameters: a connection (pointer to LibStrophe.xmpp_conn_t) and a stanza (pointer to LibStrophe.xmpp_stanza_t).

See also LibStrophe.xmpp_id_handler_add, delete_id_handler!.

Strophe.delete_id_handler! — Function

delete_id_handler!(conn, handler)

Delete an id based stanza handler.

See also LibStrophe.xmpp_id_handler_delete, add_id_handler!.

Strophe.add_timed_handler! — Function

add_timed_handler!(conn, handler, period)

Add a timed handler.

The handler will fire for the first time once the period has elapsed, and continue firing regularly after that. Strophe will try its best to fire handlers as close to the period times as it can, but accuracy will vary depending on the resolution of the event loop.

If the handler function returns true, it will be kept, and if it returns false, it will be deleted from the list of handlers.

Handlers should accept one single argument, the connection (pointer to LibStrophe.xmpp_conn_t).

See also LibStrophe.xmpp_timed_handler_add, delete_timed_handler!.

Strophe.delete_timed_handler! — Function

delete_timed_handler!(conn, handler)

Delete a timed handler.

See also LibStrophe.xmpp_timed_handler_delete, add_timed_handler!.

Strophe.add_global_timed_handler! — Function

add_global_timed_handler!([ctx,] handler, period)

Add a global timed handler.

The handler will fire for the first time once the period has elapsed, and continue firing regularly after that. Strophe will try its best to fire handlers as close to the period times as it can, but accuracy will vary depending on the resolution of the event loop.

The main difference between global and ordinary handlers:

Ordinary handler is related to a connection, fires only when the connection is in connected state and is removed once the connection is destroyed.

Global handler fires regardless of connections state and is related to a Strophe context.

The handler is executed in context of the respective event loop.

If the handler function returns true, it will be kept, and if it returns false, it will be deleted from the list of handlers.

Handlers should accept one single argument, the context (pointer to LibStrophe.xmpp_ctx_t).

See also add_timed_handler!, LibStrophe.xmpp_global_timed_handler_add, delete_global_timed_handler!.

Strophe.delete_global_timed_handler! — Function

delete_timed_handler!(ctx, handler)

Delete a global timed handler.

See also LibStrophe.xmpp_timed_handler_delete, add_global_timed_handler!.

Logging

You may be interested in the logging example.

Strophe.logger — Function

logger([_userdata::Ptr{Cvoid},] level::LibStrophe.xmpp_log_level_t, area, msg)

A logger to replace libstrophe's default logger and integrate logs in Julia's log system. _userdata is unused here and defaults to C_NULL.

The generated log event will have an area key corresponding to the area argument of this function.

See also LibStrophe.xmpp_get_default_logger.

Index

Strophe.DEFAULT_CONTEXT

Strophe.REGISTERED_HANDLERS

Strophe.sockopt_cb_keepalive

Strophe.Connection

Strophe.Context

Strophe.ParametrizedConnection

Strophe.Stanza

Strophe.StreamManagementState

Strophe.StropheError

Base.delete!

Base.getindex

Base.keys

Base.pairs

Base.setindex!

Base.values

Strophe.ClientConnection

Strophe.ComponentConnection

Strophe.add_global_timed_handler!

Strophe.add_handler!

Strophe.add_id_handler!

Strophe.add_timed_handler!

Strophe.body

Strophe.body!

Strophe.bound_jid

Strophe.child

Strophe.child!

Strophe.children

Strophe.clone!

Strophe.compressiondontreset

Strophe.compressiondontreset!

Strophe.connect

Strophe.connection

Strophe.context

Strophe.delete_global_timed_handler!

Strophe.delete_handler!

Strophe.delete_id_handler!

Strophe.delete_timed_handler!

Strophe.disablesm

Strophe.disablesm!

Strophe.disabletls

Strophe.disabletls!

Strophe.disconnect

Strophe.enablecompression

Strophe.enablecompression!

Strophe.enableflag!

Strophe.flagenabled

Strophe.flags

Strophe.flags!

Strophe.from

Strophe.from!

Strophe.id

Strophe.id!

Strophe.iq

Strophe.is_tag

Strophe.is_text

Strophe.jid

Strophe.jid!

Strophe.legacyauth

Strophe.legacyauth!

Strophe.legacyssl

Strophe.legacyssl!

Strophe.logger

Strophe.mandatorytls

Strophe.mandatorytls!

Strophe.message

Strophe.name

Strophe.name!

Strophe.ns

Strophe.ns!

Strophe.pass

Strophe.pass!

Strophe.presence

Strophe.release!

Strophe.render_to_string

Strophe.reply

Strophe.reply_error

Strophe.restore_state_management_state!

Strophe.run

Strophe.run_once

Strophe.send

Strophe.siblings

Strophe.stanza

Strophe.stop

Strophe.stream_error

Strophe.stream_management_state

Strophe.stream_management_state!

Strophe.text

Strophe.text!

Strophe.timeout!

Strophe.to

Strophe.to!

Strophe.trusttls

Strophe.trusttls!

Strophe.type

Strophe.type!