Context Objects

These functions create and manipulate Strophe context objects.

In order to support usage in a variety of environments, the Strophe library uses a runtime context object. This object contains the information on how to do memory allocation and logging. This allows the user to control how memory is allocated and what do to with log messages.

These issues do not affect programs in the common case, but many environments require special treatment. Abstracting these into a runtime context object makes it easy to use Strophe on embedded platforms.

Objects in Strophe are reference counted to ease memory management issues, but the context objects are not.

Strophe.LibStrophe.xmpp_ctx_new — Function

xmpp_ctx_new(mem, log)

Create and initialize a Strophe context object.

If mem is NULL, a default allocation setup will be used which wraps malloc(), free(), and realloc() from the standard library. If log is C_NULL, a default logger will be used which does no logging. Basic filtered logging to stderr can be done with the xmpp_get_default_logger convenience function.

Arguments

mem a pointer to an xmpp_mem_t structure or C_NULL

log a pointer to an xmpp_log_t structure or C_NULL

Returns

The allocated Strophe context object or C_NULL on an error.

Strophe.LibStrophe.xmpp_ctx_free — Function

xmpp_ctx_free(ctx)

Free a Strophe context object that is no longer in use.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

Strophe.LibStrophe.xmpp_ctx_set_verbosity — Function

xmpp_ctx_set_verbosity(ctx, level)

Set the verbosity level of a Strophe context.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

level the verbosity level

Strophe.LibStrophe.xmpp_get_default_logger — Function

xmpp_get_default_logger(level)

Get a default logger with filtering.

The default logger provides a basic logging setup which writes log messages to stderr. Only messages where level is greater than or equal to the filter level will be logged.

Arguments

level the highest level the logger will log at

Returns

The log structure for the given level

Strophe.LibStrophe.xmpp_log_level_t — Type

An enumeration for the log level used by Strophe.

Members:

XMPP_LEVEL_DEBUG

XMPP_LEVEL_INFO

XMPP_LEVEL_WARN

XMPP_LEVEL_ERROR

Strophe.LibStrophe.xmpp_ctx_t — Type

Internal, opaque, type used for storing context.

Strophe.LibStrophe.xmpp_log_handler — Type

typedef void ( * xmpp_log_handler ) ( void * userdata , xmpp_log_level_t level , const char * area , const char * msg )