Connection Management

These functions manage a connection object.

A part of those functions is listed under the SSL/TLS specific functionality section.

Strophe.LibStrophe.xmpp_sockopt_callback — Type

The function which will be called when Strophe creates a new socket.

The sock argument is a pointer that is dependent on the architecture Strophe is compiled for.

For POSIX compatible systems usage shall be: int soc = *((int*)sock);

On Windows usage shall be: SOCKET soc = *((SOCKET*)sock);

This function will be called for each socket that is created.

examples/bot.c uses a libstrophe supplied callback function that sets basic keepalive parameters (xmpp_sockopt_cb_keepalive).

examples/complex.c implements a custom function that could be useful for an application.

Connections

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

sock: A pointer to the underlying file descriptor.

Returns

0 on success, -1 on error

Strophe.LibStrophe.xmpp_conn_new — Function

xmpp_conn_new(ctx)

Create a new Strophe connection object.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

Returns

A Strophe connection object or C_NULL on an error

Strophe.LibStrophe.xmpp_conn_clone — Function

xmpp_conn_clone(conn)

Clone a Strophe connection object.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

The same conn object passed in with its reference count incremented by 1

Strophe.LibStrophe.xmpp_conn_set_sockopt_callback — Function

xmpp_conn_set_sockopt_callback(conn, callback)

Register sockopt callback Set function to be called when a new socket is created to allow setting socket options before connection is started.

If the connection is already connected, this callback will be called immediately.

To set options that can only be applied to disconnected sockets, the callback must be registered before connecting.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Strophe.LibStrophe.xmpp_conn_release — Function

xmpp_conn_release(conn)

Release a Strophe connection object.

Decrement the reference count by one for a connection, freeing the connection object if the count reaches 0.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

true if the connection object was freed and false otherwise

Strophe.LibStrophe.xmpp_conn_get_jid — Function

xmpp_conn_get_jid(conn)

Get the JID which is or will be bound to the connection.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

A string containing the full JID or C_NULL if it has not been set.

Strophe.LibStrophe.xmpp_conn_get_bound_jid — Function

xmpp_conn_get_bound_jid(conn)

Get the JID discovered during binding time.

This JID will contain the resource used by the current connection. This is useful in the case where a resource was not specified for binding.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

A string containing the full JID or C_NULL if it's not been discovered

Strophe.LibStrophe.xmpp_conn_set_jid — Function

xmpp_conn_set_jid(conn, jid)

Set the JID of the user that will be bound to the connection.

If any JID was previously set, it will be discarded. This should not be be used after a connection is created. The function will make a copy of the JID string. If the supplied JID is missing the node, SASL ANONYMOUS authentication will be used.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

jid a full or bare JID

Strophe.LibStrophe.xmpp_conn_get_pass — Function

xmpp_conn_get_pass(conn)

Get the password used for authentication of a connection.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

A string containing the password or C_NULL if it has not been set.

Strophe.LibStrophe.xmpp_conn_set_pass — Function

xmpp_conn_set_pass(conn, pass)

Set the password used to authenticate the connection.

If any password was previously set, it will be discarded. The function will make a copy of the password string.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

pass the password

Strophe.LibStrophe.xmpp_conn_get_context — Function

xmpp_conn_get_context(conn)

Get the strophe context that the connection is associated with.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

A Strophe context

Strophe.LibStrophe.xmpp_connect_client — Function

xmpp_connect_client(
    conn,
    altdomain,
    altport,
    callback,
    userdata
)

Initiate a connection to the XMPP server.

This function returns immediately after starting the connection process to the XMPP server, and notifications of connection state changes will be sent to the callback function. The domain and port to connect to are usually determined by an SRV lookup for the xmpp-client service at the domain specified in the JID. If SRV lookup fails, altdomain and altport will be used instead if specified.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

altdomain a string with domain to use if SRV lookup fails. If this is C_NULL, the domain from the JID will be used.

altport an integer port number to use if SRV lookup fails. If this is 0, the default port will be assumed.

callback a xmpp_conn_handler callback function that will receive notifications of connection status

userdata an opaque data pointer that will be passed to the callback

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_connect_component — Function

xmpp_connect_component(
    conn,
    server,
    port,
    callback,
    userdata
)

Initiate a component connection to server.

This function returns immediately after starting the connection process to the XMPP server, and notifications of connection state changes will be sent to the internal callback function that will set up handler for the component handshake as defined in XEP-0114. The domain and port to connect to must be provided in this case as the JID provided to the call serves as component identifier to the server and is not subject to DNS resolution.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

server a string with domain to use directly as the domain can't be extracted from the component name/JID. If this is not set, the call will fail. the domain from the JID will be used.

port an integer port number to use to connect to server expecting an external component. If this is 0, the port 5347 will be assumed. the default port will be assumed.

callback a xmpp_conn_handler callback function that will receive notifications of connection status

userdata an opaque data pointer that will be passed to the callback

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_connect_raw — Function

xmpp_connect_raw(
    conn,
    altdomain,
    altport,
    callback,
    userdata
)

Initiate a raw connection to the XMPP server.

Arguments and behaviour of the function are similar to xmpp_connect_client, but it skips authentication process. In opposite to xmpp_connect_client during connection process two events are generated instead of one. User's callback is called with event XMPP_CONN_RAW_CONNECT when the TCP connection with the server is established. At this point user might want to open an XMPP stream with xmpp_conn_open_stream or establish TLS session with xmpp_conn_tls_start. Event XMPP_CONN_CONNECT is generated when the XMPP stream is opened successfully and user may send stanzas over the connection.

This function doesn't use password nor node part of a jid. Therefore, the only required configuration is a domain (or full jid) passed via xmpp_conn_set_jid.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

altdomain a string with domain to use if SRV lookup fails. If this is C_NULL, the domain from the JID will be used.

altport an integer port number to use if SRV lookup fails. If this is 0, the default port will be assumed.

callback a xmpp_conn_handler callback function that will receive notifications of connection status

userdata an opaque data pointer that will be passed to the callback

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_conn_handler — Type

Pointer to a callback that will receive notifications of connection status. Defined as:

typedef void ( * xmpp_conn_handler ) ( xmpp_conn_t * conn , xmpp_conn_event_t event , int error , xmpp_stream_error_t * stream_error , void * userdata )

Strophe.LibStrophe.xmpp_conn_open_stream_default — Function

xmpp_conn_open_stream_default(conn)

Send the default opening stream tag.

The default tag is the one sent by xmpp_connect_client. User's connection handler is called with event XMPP_CONN_CONNECT when server replies with its opening tag.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Note

The connection must be connected with xmpp_connect_raw.

Strophe.LibStrophe.xmpp_conn_open_stream — Function

xmpp_conn_open_stream(conn, attributes, attributes_len)

Send an opening stream tag.

User's connection handler is called with event XMPP_CONN_CONNECT when server replies with its opening tag.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

attributes Array of strings in format: even index points to an attribute name and odd index points to its value

attributes_len Number of elements in the attributes array, it should be number of attributes multiplied by 2

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Note

The connection must be connected with xmpp_connect_raw.

Strophe.LibStrophe.xmpp_conn_tls_start — Function

xmpp_conn_tls_start(conn)

Start synchronous TLS handshake with the server.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_disconnect — Function

xmpp_disconnect(conn)

Initiate termination of the connection to the XMPP server.

This function starts the disconnection sequence by sending </stream:stream> to the XMPP server. This function will do nothing if the connection state is different from CONNECTING or CONNECTED.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Strophe.LibStrophe.xmpp_send_raw — Function

xmpp_send_raw(conn, data, len)

Send raw bytes to the XMPP server.

This function is a convenience function to send raw bytes to the XMPP server. It is used primarily by xmpp_send_raw_string. This function should be used with care as it does not validate the bytes and invalid data may result in stream termination by the XMPP server.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

data a buffer of raw bytes

len the length of the data in the buffer

Strophe.LibStrophe.xmpp_send — Function

xmpp_send(conn, stanza)

Send an XML stanza to the XMPP server.

This is the main way to send data to the XMPP server. The function will terminate without action if the connection state is not CONNECTED.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

stanza a Strophe stanza object

Strophe.LibStrophe.xmpp_conn_get_flags — Function

xmpp_conn_get_flags(conn)

Return applied flags for the connection.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Strophe.LibStrophe.xmpp_conn_set_flags — Function

xmpp_conn_set_flags(conn, flags)

Set flags for the connection.

This function applies set flags and resets unset ones. Default connection configuration is all flags unset. Flags can be applied only for a connection in disconnected state. All unsupported flags are ignored. If a flag is unset after successful set operation then the flag is not supported by current version.

Supported flags are:

LibStrophe.XMPP_CONN_FLAG_DISABLE_TLS

LibStrophe.XMPP_CONN_FLAG_MANDATORY_TLS

LibStrophe.XMPP_CONN_FLAG_LEGACY_SSL

LibStrophe.XMPP_CONN_FLAG_TRUST_TLS

LibStrophe.XMPP_CONN_FLAG_LEGACY_AUTH

LibStrophe.XMPP_CONN_FLAG_DISABLE_SM

LibStrophe.XMPP_CONN_FLAG_ENABLE_COMPRESSION

LibStrophe.XMPP_CONN_FLAG_COMPRESSION_DONT_RESET

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

fags ORed connection flags

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_conn_is_secured — Function

xmpp_conn_is_secured(conn)

Return whether TLS session is established or not.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

true if TLS session is established and false otherwise

Strophe.LibStrophe.xmpp_conn_is_connecting — Function

xmpp_conn_is_connecting(conn)

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

true if connection is in connecting state and false otherwise

Strophe.LibStrophe.xmpp_conn_is_connected — Function

xmpp_conn_is_connected(conn)

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

true if connection is established and false otherwise

Strophe.LibStrophe.xmpp_conn_is_disconnected — Function

xmpp_conn_is_disconnected(conn)

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

true if connection is in disconnected state and false otherwise

Strophe.LibStrophe.xmpp_conn_get_sm_state — Function

xmpp_conn_get_sm_state(conn)

This returns the Stream Management state of a connection object after it has been disconnected.

One can then initialise a fresh connection object and set this Stream Management state by calling xmpp_conn_set_sm_state

In case one wants to dispose of the state w/o setting it into a fresh connection object, one can call xmpp_free_sm_state

After calling this function to retrieve the state, only call one of the other two.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

The Stream Management state of the connection or C_NULL on error

Strophe.LibStrophe.xmpp_conn_set_sm_state — Function

xmpp_conn_set_sm_state(conn, sm_state)

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

sm_state A Stream Management state returned from a call to xmpp_conn_get_sm_state

Returns

XMPP_EOK (0) on success or a number less than 0 on failure.

Strophe.LibStrophe.xmpp_free_sm_state — Function

xmpp_free_sm_state(sm_state)

c.f. xmpp_conn_get_sm_state for usage documentation

Arguments

sm_state a Stream Management state returned from a call to xmpp_conn_get_sm_state

Strophe.LibStrophe.xmpp_conn_send_queue_len — Function

xmpp_conn_send_queue_len(conn)

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

The number of entries in the send queue

Strophe.LibStrophe.xmpp_conn_send_queue_drop_element — Function

xmpp_conn_send_queue_drop_element(conn, which)

Drop an element of the send queue.

This can be used to manage the send queue in case a server isn't fast enough in processing the elements you're trying to send or your outgoing bandwidth isn't fast enough to transfer everything you want to send out.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

which the element that shall be removed. See xmpp_queue_element_t

Returns

The rendered stanza. The pointer returned has to be free'd by the caller of this function.

Strophe.LibStrophe.xmpp_queue_element_t — Type

Enumeration members:

XMPP_QUEUE_OLDEST

XMPP_QUEUE_YOUNGEST

Strophe.LibStrophe.xmpp_sockopt_cb_keepalive — Function

xmpp_sockopt_cb_keepalive(conn, sock)

Example sockopt callback function An example function that can be used to set reasonable default keepalive options on sockets when registered for a connection with xmpp_conn_set_sockopt_callback

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

sock: pointer to a socket descriptor

See also xmpp_conn_set_sockopt_callback.

Strophe.LibStrophe.xmpp_conn_t — Type

Internal, opaque, type used for storing connection.

Strophe.LibStrophe.xmpp_sm_callback — Type

The function which will be called when Strophe updates its internal SM state.

Please note that you have to create a copy of the buffer, since the library will free the buffer right after return of the callback function.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

sm_state: A pointer to a buffer containing the serialized SM state.

sm_state_len: The length of sm_state.

Returns

0 on success, -1 on error

Strophe.LibStrophe.xmpp_conn_set_sm_callback — Function

This sets the Stream Management callback function

After setting this API, the library will call the given callback function each time when the internal SM state is updated.

This can be used in conjunction with xmpp_conn_restore_sm_state to e.g. implement a mechanism that retains an SM state over potential application terminations.

Arguments

conn: The Strophe connection object this callback originates from.

cb: a callback function or C_NULL to disable

ctx: The ctx pointer as passed on invocation of the callback function

Strophe.LibStrophe.xmpp_conn_restore_sm_state — Function

This restores the serialized Stream Management state

After setting this API, the library will call the given callback function each time when the internal SM state is updated.

This can be used in conjunction with xmpp_conn_set_sm_state to e.g. implement a mechanism that retains an SM state over potential application terminations.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

sm_state: a buffer as passed to the SM callback

sm_state_len: the length of sm_state

Returns

0 on success, -1 on error

Strophe.LibStrophe.xmpp_conn_event_t — Type

Status flag passed to a xmpp_conn_handler

Members:

XMPP_CONN_CONNECT

XMPP_CONN_RAW_CONNECT

XMPP_CONN_DISCONNECT

XMPP_CONN_FAIL

Strophe.LibStrophe.xmpp_stream_error_t — Type

Represents an XMPP stream error.

type::Strophe.LibStrophe.xmpp_error_type_t

text::Ptr{Int8}

stanza::Ptr{Strophe.LibStrophe._xmpp_stanza_t}

Strophe.LibStrophe.xmpp_error_type_t — Type

Error type enumeration of the error stream passed to a xmpp_conn_handler

Members:

XMPP_SE_BAD_FORMAT

XMPP_SE_BAD_NS_PREFIX

XMPP_SE_CONFLICT

XMPP_SE_CONN_TIMEOUT

XMPP_SE_HOST_GONE

XMPP_SE_HOST_UNKNOWN

XMPP_SE_IMPROPER_ADDR

XMPP_SE_INTERNAL_SERVER_ERROR

XMPP_SE_INVALID_FROM

XMPP_SE_INVALID_ID

XMPP_SE_INVALID_NS

XMPP_SE_INVALID_XML

XMPP_SE_NOT_AUTHORIZED

XMPP_SE_POLICY_VIOLATION

XMPP_SE_REMOTE_CONN_FAILED

XMPP_SE_RESOURCE_CONSTRAINT

XMPP_SE_RESTRICTED_XML

XMPP_SE_SEE_OTHER_HOST

XMPP_SE_SYSTEM_SHUTDOWN

XMPP_SE_UNDEFINED_CONDITION

XMPP_SE_UNSUPPORTED_ENCODING

XMPP_SE_UNSUPPORTED_STANZA_TYPE

XMPP_SE_UNSUPPORTED_VERSION

XMPP_SE_XML_NOT_WELL_FORMED

Strophe.LibStrophe.xmpp_sm_state_t — Type

Internal, opaque, type used for Stream-Management state storage.