SSL/TLS Specific Functionality

These functions provide SSL/TLS specific functionality.

Strophe.LibStrophe.xmpp_certfail_handler — Type

The Handler function which will be called when the TLS stack can't verify the authenticity of a Certificate that gets presented by the server we're trying to connect to.

When this function is called and details of the cert have to be kept, please copy them yourself. The cert object will be free'd automatically when this function returns.

NB: errormsg is specific per certificate on OpenSSL and the same for all certificates on GnuTLS.

Defined as:

typedef int ( * xmpp_certfail_handler ) ( const xmpp_tlscert_t * cert , const char * const errormsg )

Arguments

cert: a Strophe certificate object

errormsg: The error that caused this.

Returns

0 if the connection attempt should be terminated, 1 if the connection should be established.

Strophe.LibStrophe.xmpp_password_callback — Type

The Handler function which will be called when the TLS stack can't decrypt a password protected key file.

When this callback is called it shall write a NULL-terminated string of maximum length pw\_max - 1 to pw.

This is currently only supported for GnuTLS and OpenSSL.

On 2022-02-02 the following maximum lengths are valid:

include/gnutls/pkcs11.h:
#define GNUTLS\_PKCS11\_MAX\_PIN\_LEN 256
include/openssl/pem.h:
#define PEM\_BUFSIZE 1024

We expect the buffer to be NULL-terminated, therefore the usable lengths are:

255 for GnuTLS * 1023 for OpenSSL

Useful API's inside this callback are e.g. xmpp_conn_get_keyfile

Arguments

pw: The buffer where the password shall be stored.

pw_max: The maximum length of the password.

conn::Ptr{xmpp_conn_t}: A Strophe connection object

userdata: The userdata pointer as supplied when setting this callback.

Returns

-1 on error, else the number of bytes written to pw w/o terminating NUL byte

Strophe.LibStrophe.xmpp_cert_element_t — Type

xmpp_cert_element_t

Certificate Elements

Members:

XMPP_CERT_VERSION: X.509 Version.

XMPP_CERT_SERIALNUMBER: SerialNumber.

XMPP_CERT_SUBJECT: Subject.

XMPP_CERT_ISSUER: Issuer.

XMPP_CERT_NOTBEFORE: Issued on.

XMPP_CERT_NOTAFTER: Expires on.

XMPP_CERT_KEYALG: Public Key Algorithm.

XMPP_CERT_SIGALG: Certificate Signature Algorithm.

XMPP_CERT_FINGERPRINT_SHA1: Fingerprint SHA-1.

XMPP_CERT_FINGERPRINT_SHA256: Fingerprint SHA-256.

XMPP_CERT_ELEMENT_MAX: Last element of the enum.

Strophe.LibStrophe.xmpp_conn_set_certfail_handler — Function

xmpp_conn_set_certfail_handler(conn, hndl)

Set the Handler function which will be called when the TLS stack can't verify the CA of the server we're trying to connect to.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

hndl certfail Handler function

Strophe.LibStrophe.xmpp_conn_set_cafile — Function

xmpp_conn_set_cafile(conn, path)

Set the CAfile.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

path path to a certificate file

Strophe.LibStrophe.xmpp_conn_set_capath — Function

xmpp_conn_set_capath(conn, path)

Set the CApath.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

path path to a folder containing certificates

Strophe.LibStrophe.xmpp_conn_get_peer_cert — Function

xmpp_conn_get_peer_cert(conn)

Retrieve the peer certificate.

The returned Certificate object must be free'd by calling xmpp_tlscert_free

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

A Strophe Certificate object

Strophe.LibStrophe.xmpp_conn_set_password_callback — Function

xmpp_conn_set_password_callback(conn, cb, userdata)

Set the Callback function which will be called when the TLS stack can't decrypt a password protected key file.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

cb The callback function that shall be called

userdata An opaque data pointer that will be passed to the callback

Strophe.LibStrophe.xmpp_conn_set_password_retries — Function

xmpp_conn_set_password_retries(conn, retries)

Set the number of retry attempts to decrypt a private key file.

In case the user enters the password manually it can be useful to directly retry if the decryption of the key file failed.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

retires The number of retries that should be tried

Strophe.LibStrophe.xmpp_conn_get_keyfile — Function

xmpp_conn_get_keyfile(conn)

Retrieve the path of the key file that shall be unlocked.

This makes usually sense to be called from the xmpp_password_callback.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

A String of the path to the key file

Strophe.LibStrophe.xmpp_conn_set_client_cert — Function

xmpp_conn_set_client_cert(conn, cert, key)

Set the Client Certificate and Private Key or PKCS#12 encoded file that will be bound to the connection.

If any of them was previously set, it will be discarded. This should not be used after a connection is created. The function will make a copy of the strings passed in.

In case the Private Key is encrypted, a callback must be set via xmpp_conn_set_password_callback so the TLS stack can retrieve the password.

In case one wants to use a PKCS#12 encoded file, it should be passed via the cert parameter and key should be C_NULL. Passing a PKCS#12 file in key is deprecated.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

cert path to a certificate file or a P12 file

key path to a private key file or a P12 file

Strophe.LibStrophe.xmpp_conn_cert_xmppaddr_num — Function

xmpp_conn_cert_xmppaddr_num(conn)

Get the number of xmppAddr entries in the client certificate.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

Returns

The number of xmppAddr entries in the client certificate

Strophe.LibStrophe.xmpp_conn_cert_xmppaddr — Function

xmpp_conn_cert_xmppaddr(conn, n)

Get a specific xmppAddr entry.

Arguments

conn::Ptr{xmpp_conn_t}: A Strophe connection object

n the index of the entry, starting at 0

Returns

A string containing the xmppAddr or C_NULL if n is out of range

Strophe.LibStrophe.xmpp_tlscert_get_ctx — Function

xmpp_tlscert_get_ctx(cert)

Get the Strophe context which is assigned to this certificate.

Arguments

cert a Strophe TLS certificate object

Returns

The Strophe context object where this certificate originates from

Strophe.LibStrophe.xmpp_tlscert_get_conn — Function

xmpp_tlscert_get_conn(cert)

Get the Strophe connection which is assigned to this certificate.

Arguments

cert a Strophe TLS certificate object

Returns

The Strophe connection object where this certificate originates from

Strophe.LibStrophe.xmpp_tlscert_get_pem — Function

xmpp_tlscert_get_pem(cert)

Get the complete PEM of this certificate.

Arguments

cert a Strophe TLS certificate object

Returns

A string containing the PEM of this certificate

Strophe.LibStrophe.xmpp_tlscert_get_dnsname — Function

xmpp_tlscert_get_dnsname(cert, n)

Get the dnsName entries out of the SubjectAlternativeNames.

Note: Max. MAXNUMDNSNAMES are supported.

Arguments

cert a Strophe TLS certificate object

n which dnsName entry

Returns

A string with the n'th dnsName

Strophe.LibStrophe.xmpp_tlscert_get_string — Function

xmpp_tlscert_get_string(cert, elmnt)

Get various parts of the certificate as String.

c.f. xmpp_cert_element_t for details.

Arguments

cert a Strophe TLS certificate object

elmnt which part of the certificate

Returns

A string with the part of the certificate

Strophe.LibStrophe.xmpp_tlscert_get_description — Function

xmpp_tlscert_get_description(elmnt)

Get a descriptive string for each xmppcertelement_t.

c.f. xmpp_cert_element_t for details.

Arguments

elmnt which part of the certificate

Returns

A string with the description

Strophe.LibStrophe.xmpp_tlscert_free — Function

xmpp_tlscert_free(cert)

Free a certificate object.

Arguments

cert a Strophe TLS certificate object