Event Loop

These functions manage the Strophe event loop.

Simple tools can use Strophe.LibStrophe.xmpp_run and Strophe.LibStrophe.xmpp_stop to manage the life cycle of the program. A common idiom is to set up a few initial event handers, call Strophe.LibStrophe.xmpp_run, and then respond and react to events as they come in. At some point, one of the handlers will call Strophe.LibStrophe.xmpp_stop to quit the event loop which leads to the program terminating.

More complex programs will have their own event loops, and should ensure that Strophe.LibStrophe.xmpp_run_once is called regularly from there. For example, a GUI program will already include an event loop to process UI events from users, and Strophe.LibStrophe.xmpp_run_once would be called from an idle function.

Strophe.LibStrophe.xmpp_run_once — Function

xmpp_run_once(ctx, timeout)

Run the event loop once.

This function will run send any data that has been queued by xmpp_send and related functions and run through the Strophe even loop a single time, and will not wait more than timeout milliseconds for events. This is provided to support integration with event loops outside the library, and if used, should be called regularly to achieve low latency event handling.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

timeout time to wait for events in milliseconds

Strophe.LibStrophe.xmpp_run — Function

xmpp_run(ctx)

Start the event loop.

This function continuously calls xmpp_run_once and does not return until xmpp_stop has been called.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

Strophe.LibStrophe.xmpp_stop — Function

xmpp_stop(ctx)

Stop the event loop.

This will stop the event loop after the current iteration and cause xmpp_run to exit.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

Strophe.LibStrophe.xmpp_ctx_set_timeout — Function

xmpp_ctx_set_timeout(ctx, timeout)

Set the timeout to use when calling xmpp_run.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

timeout the time to wait for events in milliseconds