Stanza

Strophe.LibStrophe.xmpp_stanza_t — Type

Internal, opaque, type used for storing stanza.

Strophe.LibStrophe.xmpp_stanza_new — Function

xmpp_stanza_new(ctx)

Create a stanza object.

This function allocates and initializes a blank stanza object. The stanza will have a reference count of one, so the caller does not need to clone it.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

Returns

A stanza object

Strophe.LibStrophe.xmpp_stanza_clone — Function

xmpp_stanza_clone(stanza)

Clone a stanza object.

This function increments the reference count of the stanza object.

Arguments

stanza a Strophe stanza object

Returns

The stanza object with it's reference count incremented

Strophe.LibStrophe.xmpp_stanza_copy — Function

xmpp_stanza_copy(stanza)

Copy a stanza and its children.

This function copies a stanza along with all its children and returns the new stanza and children with a reference count of 1. The returned stanza will have no parent and no siblings. This function is useful for extracting a child stanza for inclusion in another tree.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

Returns

A new stanza object

Strophe.LibStrophe.xmpp_stanza_release — Function

xmpp_stanza_release(stanza)

Release a stanza object and all of its children.

This function releases a stanza object and potentially all of its children, which may cause the object(s) to be freed.

Arguments

stanza a Strophe stanza object

Returns

true if the object was freed and false otherwise

Strophe.LibStrophe.xmpp_stanza_is_text — Function

xmpp_stanza_is_text(stanza)

Determine if a stanza is a text node.

Arguments

stanza a Strophe stanza object

Returns

true if the stanza is a text node, false otherwise

Strophe.LibStrophe.xmpp_stanza_is_tag — Function

xmpp_stanza_is_tag(stanza)

Determine if a stanza is a tag node.

Arguments

stanza a Strophe stanza object

Returns

true if the stanza is a tag node, false otherwise

Strophe.LibStrophe.xmpp_stanza_to_text — Function

xmpp_stanza_to_text(stanza, buf, buflen)

Render a stanza object to text.

This function renders a given stanza object, along with its children, to text. The text is returned in an allocated, null-terminated buffer. It starts by allocating a 1024 byte buffer and reallocates more memory if that is not large enough.

Arguments

stanza a Strophe stanza object

buf a reference to a string pointer

buflen a reference to a size_t

Returns

0 on success (XMPP_EOK), and a number less than 0 on failure (XMPP_EMEM, XMPP_EINVOP)

Strophe.LibStrophe.xmpp_stanza_set_name — Function

xmpp_stanza_set_name(stanza, name)

Set the name of a stanza.

Arguments

stanza a Strophe stanza object

name a string with the name of the stanza

Returns

0 on success (XMPP_EOK), and a number less than 0 on failure (XMPP_EMEM, XMPP_EINVOP)

Strophe.LibStrophe.xmpp_stanza_get_name — Function

xmpp_stanza_get_name(stanza)

Get the stanza name.

This function returns a pointer to the stanza name. If the caller needs to store this data, it must make a copy.

Arguments

stanza a Strophe stanza object

Returns

A string with the stanza name

Strophe.LibStrophe.xmpp_stanza_get_attribute_count — Function

xmpp_stanza_get_attribute_count(stanza)

Count the attributes in a stanza object.

Arguments

stanza a Strophe stanza object

Returns

The number of attributes for the stanza object

Strophe.LibStrophe.xmpp_stanza_get_attributes — Function

xmpp_stanza_get_attributes(stanza, attr, attrlen)

Get all attributes for a stanza object.

This function populates the array with attributes from the stanza. The attr array will be in the format: attr[i] = attribute name, attr[i+1] = attribute value.

Arguments

stanza a Strophe stanza object

attr the string array to populate

attrlen the size of the array

Returns

The number of slots used in the array, which will be 2 times the number of attributes in the stanza

Strophe.LibStrophe.xmpp_stanza_set_attribute — Function

xmpp_stanza_set_attribute(stanza, key, value)

Set an attribute for a stanza object.

Arguments

stanza a Strophe stanza object

key a string with the attribute name

value a string with the attribute value

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_stanza_set_ns — Function

xmpp_stanza_set_ns(stanza, ns)

Set the stanza namespace.

This is a convenience function equivalent to calling: xmpp_stanza_set_attribute(stanza, "xmlns", ns)

Arguments

stanza a Strophe stanza object

ns a string with the namespace

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_stanza_add_child_ex — Function

xmpp_stanza_add_child_ex(stanza, child, do_clone)

Add a child stanza to a stanza object.

If do_clone is 1, user keeps reference to the child stanza and must call xmpp_stanza_release to release the reference. If do_clone is 0, user transfers ownership and must not neither call xmpp_stanza_release for the child stanza nor use it.

Arguments

stanza a Strophe stanza object

child the child stanza object

do_clone 1 to increase ref count of child (default for xmpp_stanza_add_child)

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_stanza_add_child — Function

xmpp_stanza_add_child(stanza, child)

Add a child stanza to a stanza object.

This function clones the child and appends it to the stanza object's children.

Arguments

stanza a Strophe stanza object

child the child stanza object

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_stanza_set_text — Function

xmpp_stanza_set_text(stanza, text)

Set the text data for a text stanza.

This function copies the text given and sets the stanza object's text to it. Attempting to use this function on a stanza that has a name will fail with XMPP_EINVOP. This function takes the text as a null-terminated string.

Arguments

stanza a Strophe stanza object

text a string with the text

Returns

0 on success (XMPP_EOK), and a number less than 0 on failure

Strophe.LibStrophe.xmpp_stanza_set_text_with_size — Function

xmpp_stanza_set_text_with_size(stanza, text, size)

Set the text data for a text stanza.

This function copies the text given and sets the stanza object's text to it. Attempting to use this function on a stanza that has a name will fail with XMPP_EINVOP. This function takes the text as buffer and a length as opposed to a null-terminated string.

Arguments

stanza a Strophe stanza object

text a buffer with the text

size the length of the text

Returns

0 on success (XMPP_EOK), and a number less than 0 on failure

Strophe.LibStrophe.xmpp_stanza_get_id — Function

xmpp_stanza_get_id(stanza)

Get the 'id' attribute of the stanza object.

This is a convenience function equivalent to: xmpp_stanza_get_attribute(stanza, "id")

Arguments

stanza a Strophe stanza object

Returns

A string with the 'id' attribute value

Strophe.LibStrophe.xmpp_stanza_get_ns — Function

xmpp_stanza_get_ns(stanza)

Get the namespace attribute of the stanza object.

This is a convenience function equivalent to: xmpp_stanza_get_attribute(stanza, "xmlns")

Arguments

stanza a Strophe stanza object

Returns

A string with the 'xmlns' attribute value

Strophe.LibStrophe.xmpp_stanza_get_type — Function

xmpp_stanza_get_type(stanza)

Get the 'type' attribute of the stanza object.

This is a convenience function equivalent to: xmpp_stanza_get_attribute(stanza, "type")

Arguments

stanza a Strophe stanza object

Returns

A string with the 'type' attribute value

Strophe.LibStrophe.xmpp_stanza_get_to — Function

xmpp_stanza_get_to(stanza)

Get the 'to' attribute of the stanza object.

This is a convenience function equivalent to: xmpp_stanza_get_attribute(stanza, "to")

Arguments

stanza a Strophe stanza object

Returns

A string with the 'to' attribute value

Strophe.LibStrophe.xmpp_stanza_get_from — Function

xmpp_stanza_get_from(stanza)

Get the 'from' attribute of the stanza object.

This is a convenience function equivalent to: xmpp_stanza_get_attribute(stanza, "from")

Arguments

stanza a Strophe stanza object

Returns

A string with the 'from' attribute value

Strophe.LibStrophe.xmpp_stanza_get_child_by_name — Function

xmpp_stanza_get_child_by_name(stanza, name)

Get the first child of stanza with name.

This function searches all the immediate children of stanza for a child stanza that matches the name. The first matching child is returned.

Arguments

stanza a Strophe stanza object

name a string with the name to match

Returns

The matching child stanza object or C_NULL if no match was found

Strophe.LibStrophe.xmpp_stanza_get_child_by_ns — Function

xmpp_stanza_get_child_by_ns(stanza, ns)

Get the first child of a stanza with a given namespace.

This function searches all the immediate children of a stanza for a child stanza that matches the namespace provided. The first matching child is returned.

Arguments

stanza a Strophe stanza object

ns a string with the namespace to match

Returns

The matching child stanza object or C_NULL if no match was found

Strophe.LibStrophe.xmpp_stanza_get_child_by_name_and_ns — Function

xmpp_stanza_get_child_by_name_and_ns(stanza, name, ns)

Get the first child of stanza with name and a given namespace.

This function searches all the immediate children of stanza for a child stanza that matches the name and namespace provided. The first matching child is returned.

Arguments

stanza a Strophe stanza object

name a string with the name to match

ns a string with the namespace to match

Returns

The matching child stanza object or C_NULL if no match was found

Strophe.LibStrophe.xmpp_stanza_get_children — Function

xmpp_stanza_get_children(stanza)

Get the list of children.

This function returns the first child of the stanza object. The rest of the children can be obtained by calling xmpp_stanza_get_next to iterate over the siblings.

Arguments

stanza a Strophe stanza object

Returns

The first child stanza or C_NULL if there are no children

Strophe.LibStrophe.xmpp_stanza_get_next — Function

xmpp_stanza_get_next(stanza)

Get the next sibling of a stanza.

Arguments

stanza a Strophe stanza object

Returns

The next sibling stanza or C_NULL if there are no more siblings

Strophe.LibStrophe.xmpp_stanza_get_text — Function

xmpp_stanza_get_text(stanza)

Get the text data for a text stanza.

This function copies the text data from a stanza and returns the new allocated string. The caller is responsible for freeing this string with xmpp_free.

Arguments

stanza a Strophe stanza object

Returns

An allocated string with the text data

Strophe.LibStrophe.xmpp_stanza_get_text_ptr — Function

xmpp_stanza_get_text_ptr(stanza)

Get the text data pointer for a text stanza.

This function copies returns the raw pointer to the text data in the stanza. This should only be used in very special cases where the caller needs to translate the datatype as this will save a double allocation. The caller should not hold onto this pointer, and is responsible for allocating a copy if it needs one.

Arguments

stanza a Strophe stanza object

Returns

An string pointer to the data or C_NULL

Strophe.LibStrophe.xmpp_stanza_set_id — Function

xmpp_stanza_set_id(stanza, id)

Set the 'id' attribute of a stanza.

This is a convenience function for: xmpp_stanza_set_attribute(stanza, 'id', id)

Arguments

stanza a Strophe stanza object

id a string containing the 'id' value

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_stanza_set_type — Function

xmpp_stanza_set_type(stanza, type)

Set the 'type' attribute of a stanza.

This is a convenience function for: xmpp_stanza_set_attribute(stanza, 'type', type)

Arguments

stanza a Strophe stanza object

type a string containing the 'type' value

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_stanza_set_to — Function

xmpp_stanza_set_to(stanza, to)

Set the 'to' attribute of a stanza.

This is a convenience function for: xmpp_stanza_set_attribute(stanza, 'to', to)

Arguments

stanza a Strophe stanza object

to a string containing the 'to' value

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_stanza_set_from — Function

xmpp_stanza_set_from(stanza, from)

Set the 'from' attribute of a stanza.

This is a convenience function for: xmpp_stanza_set_attribute(stanza, 'from', from)

Arguments

stanza a Strophe stanza object

from a string containing the 'from' value

Returns

XMPP_EOK (0) on success or a number less than 0 on failure

Strophe.LibStrophe.xmpp_stanza_get_attribute — Function

xmpp_stanza_get_attribute(stanza, name)

Get an attribute from a stanza.

This function returns a pointer to the attribute value. If the caller wishes to save this value it must make its own copy.

Arguments

stanza a Strophe stanza object

name a string containing attribute name

Returns

A string with the attribute value or C_NULL on an error

Strophe.LibStrophe.xmpp_stanza_del_attribute — Function

xmpp_stanza_del_attribute(stanza, name)

Delete an attribute from a stanza.

Arguments

stanza a Strophe stanza object

name a string containing attribute name

Returns

0 on success (XMPP_EOK), and a number less than 0 on failure

Strophe.LibStrophe.xmpp_stanza_reply — Function

xmpp_stanza_reply(stanza)

Create a stanza object in reply to another.

This function makes a copy of a stanza object with the attribute "to" set its original "from". The stanza will have a reference count of one, so the caller does not need to clone it.

Arguments

stanza a Strophe stanza object

Returns

A new Strophe stanza object

Strophe.LibStrophe.xmpp_stanza_reply_error — Function

xmpp_stanza_reply_error(stanza, error_type, condition, text)

Create an error stanza in reply to the provided stanza.

Check https://tools.ietf.org/html/rfc6120#section-8.3 for details.

Arguments

stanza a Strophe stanza object

error_type type attribute in the <error/> child element

condition the defined-condition (e.g. "item-not-found")

text optional description, may be C_NULL

Returns

A new Strophe stanza object

Strophe.LibStrophe.xmpp_message_new — Function

xmpp_message_new(ctx, type, to, id)

Create a <message/> stanza object with given attributes.

Attributes are optional and may be C_NULL.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

type attribute 'type'

to attribute 'to'

id attribute 'id'

Returns

A new Strophe stanza object

Strophe.LibStrophe.xmpp_message_get_body — Function

xmpp_message_get_body(msg)

Get text from <body/> child element.

This function returns new allocated string. The caller is responsible for freeing this string with xmpp_free.

Arguments

msg well formed <message/> stanza

Returns

Allocated string or C_NULL on failure (no <body/> element or memory allocation error)

Strophe.LibStrophe.xmpp_message_set_body — Function

xmpp_message_set_body(msg, text)

Add <body/> child element to a <message/> stanza with the given text.

Arguments

msg a <message> stanza object without <body/> child element.

text The text that shall be placed in the body.

Returns

0 on success (XMPP_EOK), and a number less than 0 on failure (XMPP_EMEM, XMPP_EINVOP)

Strophe.LibStrophe.xmpp_iq_new — Function

xmpp_iq_new(ctx, type, id)

Create an <iq/> stanza object with given attributes.

Attributes are optional and may be C_NULL.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

type attribute 'type'

id attribute 'id'

Returns

A new Strophe stanza object

Strophe.LibStrophe.xmpp_presence_new — Function

xmpp_presence_new(ctx)

Create a <presence/> stanza object.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

Returns

A new Strophe stanza object

Strophe.LibStrophe.xmpp_error_new — Function

xmpp_error_new(ctx, type, text)

Create an <stream:error/> stanza object with given type and error text.

The error text is optional and may be C_NULL.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

type enum of stropheerrortype_t

text content of a 'text'

Returns

A new Strophe stanza object

Strophe.LibStrophe.xmpp_stanza_new_from_string — Function

xmpp_stanza_new_from_string(ctx, str)

Create a stanza object from the string.

This function allocates and initializes a stanza object which represents stanza located in the string. The stanza will have a reference count of one, so the caller does not need to clone it.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

str stanza in NULL terminated string representation

Returns

A stanza object or C_NULL on an error