Encodings

Strophe.LibStrophe.xmpp_base64_encode — Function

xmpp_base64_encode(ctx, data, len)

Base64 encoding routine.

Returns

An allocated string which must be freed with xmpp_free.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

data buffer to encode

len size of the data buffer

Returns

An allocated null-terminated string or C_NULL on error

Strophe.LibStrophe.xmpp_base64_decode_str — Function

xmpp_base64_decode_str(ctx, base64, len)

Base64 decoding routine.

Returns

An allocated string which must be freed with xmpp_free. User calls this function when the result must be a string. When decoded buffer contains ' ' NULL is returned.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

base64 encoded buffer

len size of the buffer

Returns

An allocated null-terminated string or C_NULL on error

Strophe.LibStrophe.xmpp_base64_decode_bin — Function

xmpp_base64_decode_bin(ctx, base64, len, out, outlen)

Base64 decoding routine.

Returns

An allocated buffer which must be freed with xmpp_free.

Arguments

ctx::Ptr{xmpp_ctx_t}: A Strophe context object

base64 encoded buffer

len size of the encoded buffer

out allocated buffer is stored here

outlen size of the allocated buffer

Note

on an error the out pointer will be C_NULL