Constants

Return Values

Strophe.LibStrophe.XMPP_EOK — Constant

Return code signaling no error

Strophe.LibStrophe.XMPP_EMEM — Constant

Memory related failure error code.

This is returned on allocation errors and signals that the host may be out of memory.

Strophe.LibStrophe.XMPP_EINVOP — Constant

Invalid operation error code.

This error code is returned when the operation was invalid and signals that the Strophe API is being used incorrectly.

Strophe.LibStrophe.XMPP_EINT — Constant

Internal failure error code.

Namespaces

Strophe.LibStrophe.XMPP_NS_CLIENT — Constant

jabber:client namespace string.

Strophe.LibStrophe.XMPP_NS_COMPONENT — Constant

jabber:component:accept namespace string.

Strophe.LibStrophe.XMPP_NS_STREAMS — Constant

http://etherx.jabber.org/streams namespace string.

Strophe.LibStrophe.XMPP_NS_STREAMS_IETF — Constant

urn:ietf:params:xml:ns:xmpp-streams namespace string.

Strophe.LibStrophe.XMPP_NS_STANZAS_IETF — Constant

urn:ietf:params:xml:ns:xmpp-stanzas namespace string.

Strophe.LibStrophe.XMPP_NS_TLS — Constant

urn:ietf:params:xml:ns:xmpp-tls namespace string.

Strophe.LibStrophe.XMPP_NS_SASL — Constant

urn:ietf:params:xml:ns:xmpp-sasl namespace string.

Strophe.LibStrophe.XMPP_NS_BIND — Constant

urn:ietf:params:xml:ns:xmpp-bind namespace string.

Strophe.LibStrophe.XMPP_NS_SESSION — Constant

urn:ietf:params:xml:ns:xmpp-session namespace string.

Strophe.LibStrophe.XMPP_NS_AUTH — Constant

jabber:iq:auth namespace string.

Strophe.LibStrophe.XMPP_NS_DISCO_INFO — Constant

http://jabber.org/protocol/disco#info namespace string.

Strophe.LibStrophe.XMPP_NS_DISCO_ITEMS — Constant

http://jabber.org/protocol/disco#items namespace string.

Strophe.LibStrophe.XMPP_NS_ROSTER — Constant

jabber:iq:roster namespace string.

Strophe.LibStrophe.XMPP_NS_REGISTER — Constant

jabber:iq:register namespace string.

Strophe.LibStrophe.XMPP_NS_SM — Constant

urn:xmpp:sm:3 namespace string.

Strophe.LibStrophe.XMPP_NS_COMPRESSION — Constant

http://jabber.org/protocol/compress namespace string.

Strophe.LibStrophe.XMPP_NS_FEATURE_COMPRESSION — Constant

http://jabber.org/features/compress namespace string.

Connection Flags

Strophe.LibStrophe.XMPP_CONN_FLAG_DISABLE_TLS — Constant

Disable TLS.

Strophe.LibStrophe.XMPP_CONN_FLAG_MANDATORY_TLS — Constant

Make TLS mandatory.

Strophe.LibStrophe.XMPP_CONN_FLAG_LEGACY_SSL — Constant

Enable legacy SSL

Strophe.LibStrophe.XMPP_CONN_FLAG_TRUST_TLS — Constant

Trust TLS certificate.

Strophe.LibStrophe.XMPP_CONN_FLAG_LEGACY_AUTH — Constant

Enable legacy auth.

Strophe.LibStrophe.XMPP_CONN_FLAG_DISABLE_SM — Constant

Disable stream management.

Strophe.LibStrophe.XMPP_CONN_FLAG_ENABLE_COMPRESSION — Constant

Enable compression.

Strophe.LibStrophe.XMPP_CONN_FLAG_COMPRESSION_DONT_RESET — Constant

Do not reset compression.